Add validator on Delete endpoint & Sample requests

+++ HogwartsApiRest/Controllers/StudentsController.cs
@@ -16,7 +16,7 @@ namespace HogwartsApiRest.Controllers
         }
 
         /// <summary>
-        /// Get all student entries in the data base.
+        /// Gets all students in the data base.
         /// </summary>
         /// <response code="404">Something went wrong.</response>              
         [HttpGet]
@@ -26,7 +26,7 @@ namespace HogwartsApiRest.Controllers
         }
 
         /// <summary>
-        /// Get a specific student given the id.
+        /// Gets a specific student given the id.
         /// </summary>              
         /// <response code="404">No student found with the given id.</response>
         [HttpGet]
@@ -42,9 +42,21 @@ namespace HogwartsApiRest.Controllers
         }
 
         /// <summary>
-        /// Create a new student entry.
-        /// </summary>              
-        /// <response code="201">New student entry created successfully.</response>
+        /// Creates a new student.
+        /// </summary>
+        /// <remarks>
+        /// Sample request:
+        ///
+        ///     {
+        ///       "houseId": 1,
+        ///       "name": "Harry",
+        ///       "lastName": "Potter",
+        ///       "identification": 421515214,
+        ///       "age": 20
+        ///     }
+        ///
+        /// </remarks>
+        /// <response code="201">New student created successfully.</response>
         /// <response code="400">Some fields do not meet the requested format.</response>
         [HttpPost]
         public async Task<IActionResult> Post([FromBody] DtoStudentRequest request)
@@ -61,9 +73,21 @@ namespace HogwartsApiRest.Controllers
         }
 
         /// <summary>
-        /// Update an existing student entry.
-        /// </summary>              
-        /// <response code="202">Student entry updated successfully.</response>
+        /// Updates an existing student.
+        /// </summary>
+        /// <remarks>
+        /// Sample request:
+        ///
+        ///     {
+        ///       "houseId": 1,
+        ///       "name": "Ron",
+        ///       "lastName": "Weasley",
+        ///       "identification": 425621412,
+        ///       "age": 20
+        ///     }
+        ///
+        /// </remarks>
+        /// <response code="202">Student updated successfully.</response>
         /// <response code="400">Some fields do not meet the requested format.</response>
         [HttpPut]
         [Route("{id:int}")]
@@ -81,15 +105,21 @@ namespace HogwartsApiRest.Controllers
         }
 
         /// <summary>
-        /// Deletes a specific student entry.
+        /// Deletes a specific student.
         /// </summary>
         /// <param name="id"></param>        
         /// <response code="202">Student deleted with the given id.</response>
+        /// <response code="404">No student found with the given id.</response>
         [ProducesResponseType(StatusCodes.Status202Accepted)]
         [HttpDelete]
         [Route("{id:int}")]
         public async Task<IActionResult> Delete(int id)
         {
+            var response = await _service.GetByIdAsync(id);
+
+            if (response == null)
+                return NotFound();
+
             await _service.DeleteAsync(id);
 
             return Accepted(new
